Add Coveralls; Seperate TXN Construction
