+++ .github/workflows/deployment.yml
@@ -45,3 +45,8 @@ jobs:
 
     - name: Coverage
       uses: codecov/codecov-action@v1
+
+    - name: Coveralls GitHub Action
+      uses: coverallsapp/github-action@v1.1.2
+      with:
+        github-token: ${{ secrets.GITHUB_TOKEN }}
